+++ src/main/java/com/programmers/voucher/ConsoleMenu.java
@@ -6,7 +6,7 @@ import com.programmers.voucher.domain.voucher.controller.VoucherController;
 import com.programmers.voucher.domain.voucher.domain.Voucher;
 import com.programmers.voucher.global.io.ConsoleCommandType;
 import com.programmers.voucher.global.io.Console;
-import com.programmers.voucher.domain.voucher.request.VoucherCreateRequest;
+import com.programmers.voucher.domain.voucher.dto.request.VoucherCreateRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
