+++ src/main/java/com/programmers/voucher/domain/voucher/controller/VoucherController.java
@@ -1,7 +1,7 @@
 package com.programmers.voucher.domain.voucher.controller;
 
 import com.programmers.voucher.domain.voucher.domain.Voucher;
-import com.programmers.voucher.domain.voucher.request.VoucherCreateRequest;
+import com.programmers.voucher.domain.voucher.dto.request.VoucherCreateRequest;
 import com.programmers.voucher.domain.voucher.service.VoucherService;
 import org.springframework.stereotype.Controller;
